Create main.cpp
This is exactly what I had submitted for my assignment when I took the class. I hope to make a program later that will automatically enter in the borders so that it doesn't have to be hard codded (which looks very messy). Look forward to clean ups in the future.